requirements.txt
discord.py
pyinstaller
